separate map constants file
